Rename HelperLogbook -> Logbook
Use the same name for file and component.
